refactor(decompose): split webapp/e2e-tests/electron/for_feature_development_not_LT_verification/graph/perf/electron-agent-storm-perf.spec.ts into folder
