+++ webapp/e2e-tests/electron/for_feature_development_not_LT_verification/graph/perf/electron-agent-storm-perf/fakeAgents.ts
@@ -1,0 +1,73 @@
+import * as path from 'node:path'
+import { existsSync } from 'node:fs'
+import { createRequire } from 'node:module'
+
+import { agentRuntime } from '@vt/agent-runtime'
+
+const require = createRequire(import.meta.url)
+
+export interface AgentResult {
+    readonly terminalId: string
+    readonly spawnSuccess: boolean
+    readonly startedAtMs: number
+    readonly exitedAtMs: number | null
+    readonly exitCode: number | null
+    readonly stdoutSnippet: string
+    readonly errorMessage?: string
+}
+
+export function buildFakeAgentScript(nodesPerAgent: number): object {
+    const actions: object[] = []
+    for (let i = 0; i < nodesPerAgent; i++) {
+        actions.push({
+            type: 'create_node',
+            title: `Perf E2E Node ${i}`,
+            summary: `Synthetic node ${i} produced by e2e perf-agent-storm.`,
+            content: `Node body for index ${i}. Generated by the e2e perf harness.`,
+        })
+    }
+    actions.push({ type: 'exit', code: 0 })
+    return { actions }
+}
+
+export function buildAgentPrompt(script: object): string {
+    return `### FAKE_AGENT_SCRIPT ###\n${JSON.stringify(script)}\n### END_FAKE_AGENT_SCRIPT ###`
+}
+
+function resolveRepoRoot(specDir: string): string {
+    // webapp/e2e-tests/.../graph/perf/electron-agent-storm-perf.spec.ts
+    // -> ../../../../../.. is the worktree root from the spec directory.
+    return path.resolve(specDir, '..', '..', '..', '..', '..', '..')
+}
+
+export function resolveFakeAgentEntrypoint(specDir: string): { dir: string; entry: string } {
+    const repoRoot = resolveRepoRoot(specDir)
+    const dir = path.join(repoRoot, 'tools', 'vt-fake-agent')
+    const entry = path.join(dir, 'src', 'index.ts')
+    if (!existsSync(entry)) throw new Error(`vt-fake-agent entrypoint not found at ${entry}`)
+    return { dir, entry }
+}
+
+export function resolveTsxImportPath(): string {
+    return require.resolve('tsx')
+}
+
+export async function waitForExit(
+    terminalId: string,
+    exitedTerminals: Map<string, { code: number; atMs: number }>,
+    timeoutMs: number,
+): Promise<{ code: number; atMs: number } | null> {
+    const deadline = Date.now() + timeoutMs
+    while (Date.now() < deadline) {
+        const found = exitedTerminals.get(terminalId)
+        if (found) return found
+        const rec = agentRuntime.getTerminalRecords().find(r => r.terminalId === terminalId)
+        if (rec?.status === 'exited') {
+            const entry = { code: 0, atMs: Date.now() }
+            exitedTerminals.set(terminalId, entry)
+            return entry
+        }
+        await new Promise(r => setTimeout(r, 250))
+    }
+    return null
+}
